backend/requirements.txt
# PharmaInsight Pro - Backend Dependencies

# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Data Processing
pandas==2.1.4
numpy==1.26.2
openpyxl==3.1.2

# Forecasting
scikit-learn==1.3.2
statsmodels==0.14.1

# CORS & Security
python-dotenv==1.0.0

# Date handling
python-dateutil==2.8.2

# HTTP Client for AI API
httpx==0.26.0
